Refactoring class DefaultSteps. Class DefaultSteps is divided into 8 new classes with steps and 1 class with base methods Delete duplicate step @Тогда("^список \"([^\"]*)\" состоит из элементов из таблицы$”) Made one step of two @Тогда("^(?:ссылка|кнопка|поле|элемент) \"([^\"]*)\" (?:недоступно|недоступен|недоступна)$”) Unit tests is devided into 6 new classes
