--- src/main/java/ru/alfabank/steps/DefaultSteps.java
@@ -1,1199 +1,0 @@
-/**
- * Copyright 2017 Alfa Laboratory
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package ru.alfabank.steps;
-
-import com.codeborne.selenide.Condition;
-import com.codeborne.selenide.Selenide;
-import com.codeborne.selenide.SelenideElement;
-import cucumber.api.DataTable;
-import cucumber.api.java.ru.*;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.RandomStringUtils;
-import org.hamcrest.Matchers;
-import org.openqa.selenium.*;
-import org.openqa.selenium.Dimension;
-import org.openqa.selenium.interactions.Actions;
-import ru.alfabank.alfatest.cucumber.api.AkitaScenario;
-
-import java.awt.*;
-import java.awt.datatransfer.Clipboard;
-import java.awt.datatransfer.ClipboardOwner;
-import java.awt.datatransfer.StringSelection;
-import java.io.File;
-import java.text.DecimalFormat;
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Random;
-import java.util.concurrent.ThreadLocalRandom;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-
-import static com.codeborne.selenide.Condition.not;
-import static com.codeborne.selenide.Selenide.*;
-import static com.codeborne.selenide.WebDriverRunner.*;
-import static java.lang.String.format;
-import static java.util.Objects.isNull;
-import static java.util.stream.Collectors.toList;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.*;
-import static org.junit.Assert.*;
-import static ru.alfabank.alfatest.cucumber.ScopedVariables.resolveVars;
-import static ru.alfabank.tests.core.helpers.PropertyLoader.*;
-
-/**
- * В akitaScenario используется хранилище переменных. Для сохранения/изъятия переменных используются методы setVar/getVar
- * Каждая страница, с которой предполагается взаимодействие, должна быть описана в соответствующем классе,
- * наследующем AkitaPage. Для каждого элемента следует задать имя на русском, через аннотацию @Name, чтобы искать
- * можно было именно по русскому описанию, а не по селектору. Селекторы следует хранить только в классе страницы,
- * не в степах, в степах - взаимодействие по русскому названию элемента.
- */
-@Slf4j
-public class DefaultSteps {
-
-    private AkitaScenario akitaScenario = AkitaScenario.getInstance();
-
-    private static final int DEFAULT_TIMEOUT = loadPropertyInt("waitingCustomElementsTimeout", 10000);
-
-    /**
-     * Значение заданной переменной из application.properties сохраняется в переменную в akitaScenario
-     * для дальнейшего использования
-     */
-    @И("^сохранено значение \"([^\"]*)\" из property файла в переменную \"([^\"]*)\"$")
-    public void saveValueToVar(String propertyVariableName, String variableName) {
-        propertyVariableName = loadProperty(propertyVariableName);
-        akitaScenario.setVar(variableName, propertyVariableName);
-        akitaScenario.write("Значение сохраненной переменной " + propertyVariableName);
-    }
-
-    /**
-     * Выполняется обновление страницы
-     */
-    @И("^выполнено обновление текущей страницы$")
-    public void refreshPage() {
-        refresh();
-    }
-
-    /**
-     * Выполняется переход по заданной ссылке,
-     * ссылка берется из property / переменной, если такая переменная не найдена,
-     * то берется переданное значение
-     * при этом все ключи переменных в фигурных скобках
-     * меняются на их значения из хранилища akitaScenario
-     */
-    @Когда("^совершен переход по ссылке \"([^\"]*)\"$")
-    public void goToUrl(String address) {
-        String url = resolveVars(getPropertyOrStringVariableOrValue(address));
-        open(url);
-        akitaScenario.write("Url = " + url);
-    }
-
-    /**
-     * Проверка, что текущий URL совпадает с ожидаемым
-     * (берется из property / переменной, если такая переменная не найдена,
-     * то берется переданное значение)
-     */
-    @Тогда("^текущий URL равен \"([^\"]*)\"$")
-    public void checkCurrentURL(String url) {
-        String currentUrl = url();
-        String expectedUrl = resolveVars(getPropertyOrStringVariableOrValue(url));
-        assertThat("Текущий URL не совпадает с ожидаемым", currentUrl, is(expectedUrl));
-    }
-
-    /**
-     * Проверка, что текущий URL не совпадает с ожидаемым
-     * (берется из property / переменной, если такая переменная не найдена,
-     * то берется переданное значение)
-     */
-    @Тогда("^текущий URL не равен \"([^\"]*)\"$")
-    public void checkCurrentURLIsNotEquals(String url) {
-        String currentUrl = url();
-        String expectedUrl = resolveVars(getPropertyOrStringVariableOrValue(url));
-        assertThat("Текущий URL совпадает с ожидаемым", currentUrl, Matchers.not(expectedUrl));
-    }
-
-    /**
-     * На странице происходит клик по заданному элементу
-     */
-    @И("^выполнено нажатие на (?:кнопку|поле|блок) \"([^\"]*)\"$")
-    public void clickOnElement(String elementName) {
-        akitaScenario.getCurrentPage().getElement(elementName).click();
-    }
-
-    /**
-     * Проверка появления элемента(не списка) на странице в течение DEFAULT_TIMEOUT.
-     * В случае, если свойство "waitingCustomElementsTimeout" в application.properties не задано,
-     * таймаут равен 10 секундам
-     */
-    @Тогда("^элемент \"([^\"]*)\" отображается на странице$")
-    public void elemIsPresentedOnPage(String elementName) {
-        akitaScenario.getCurrentPage().waitElementsUntil(
-            Condition.appear, DEFAULT_TIMEOUT, akitaScenario.getCurrentPage().getElement(elementName)
-        );
-    }
-
-    /**
-     * Проверка появления элемента(не списка) на странице в течение
-     * заданного количества секунд
-     */
-    @Тогда("^элемент \"([^\"]*)\" отобразился на странице в течение (\\d+) (?:секунд|секунды)")
-    public void testElementAppeared(String elementName, int seconds) {
-        akitaScenario.getCurrentPage().waitElementsUntil(
-            Condition.appear, seconds * 1000, akitaScenario.getCurrentPage().getElement(elementName)
-        );
-    }
-
-    /**
-     * Проверка появления списка на странице в течение DEFAULT_TIMEOUT.
-     * В случае, если свойство "waitingCustomElementsTimeout" в application.properties не задано,
-     * таймаут равен 10 секундам
-     */
-    @Тогда("^список \"([^\"]*)\" отображается на странице$")
-    public void listIsPresentedOnPage(String elementName) {
-        akitaScenario.getCurrentPage().waitElementsUntil(
-            Condition.appear, DEFAULT_TIMEOUT, akitaScenario.getCurrentPage().getElementsList(elementName)
-        );
-    }
-
-    /**
-     * Проверка того, что элемент исчезнет со страницы (станет невидимым) в течение DEFAULT_TIMEOUT.
-     * В случае, если свойство "waitingCustomElementsTimeout" в application.properties не задано,
-     * таймаут равен 10 секундам
-     */
-    @Тогда("^ожидается исчезновение элемента \"([^\"]*)\"")
-    public void elemDisappered(String elementName) {
-        akitaScenario.getCurrentPage().waitElementsUntil(
-            Condition.disappears, DEFAULT_TIMEOUT, akitaScenario.getCurrentPage().getElement(elementName));
-    }
-
-    /**
-     * Проверка того, что все элементы, которые описаны в классе страницы с аннотацией @Name,
-     * но без аннотации @Optional появились на странице
-     * в течение WAITING_APPEAR_TIMEOUT, которое равно значению свойства "waitingAppearTimeout"
-     * из application.properties. Если свойство не найдено, время таймаута равно 8 секундам
-     */
-    @Тогда("^(?:страница|блок|форма|вкладка) \"([^\"]*)\" (?:загрузилась|загрузился)$")
-    public void loadPage(String nameOfPage) {
-        akitaScenario.setCurrentPage(akitaScenario.getPage(nameOfPage));
-        if (isIE()) {
-            akitaScenario.getCurrentPage().ieAppeared();
-        } else akitaScenario.getCurrentPage().appeared();
-    }
-
-    /**
-     * Проверка того, что все элементы, которые описаны в классе страницы с аннотацией @Name,
-     * но без аннотации @Optional, не появились на странице
-     */
-    @Тогда("^(?:страница|блок|форма|вкладка) \"([^\"]*)\" не (?:загрузилась|загрузился)$")
-    public void loadPageFailed(String nameOfPage) {
-        akitaScenario.setCurrentPage(akitaScenario.getPage(nameOfPage));
-        if (isIE()) {
-            akitaScenario.getCurrentPage().ieDisappeared();
-        } else akitaScenario.getCurrentPage().disappeared();
-    }
-
-    /**
-     * Устанавливается значение переменной в хранилище переменных. Один из кейсов: установка login пользователя
-     */
-    @И("^установлено значение переменной \"([^\"]*)\" равным \"(.*)\"$")
-    public void setVariable(String variableName, String value) {
-        value = getPropertyOrValue(value);
-        akitaScenario.setVar(variableName, value);
-    }
-
-    /**
-     * Проверка равенства двух переменных из хранилища
-     */
-    @Тогда("^значения в переменных \"([^\"]*)\" и \"([^\"]*)\" совпадают$")
-    public void compareTwoVariables(String firstVariableName, String secondVariableName) {
-        String firstValueToCompare = akitaScenario.getVar(firstVariableName).toString();
-        String secondValueToCompare = akitaScenario.getVar(secondVariableName).toString();
-        assertThat(String.format("Значения в переменных [%s] и [%s] не совпадают", firstVariableName, secondVariableName),
-            firstValueToCompare, equalTo(secondValueToCompare));
-    }
-
-    /**
-     * Проверка неравенства двух переменных из хранилища
-     */
-    @Тогда("^значения в переменных \"([^\"]*)\" и \"([^\"]*)\" не совпадают$")
-    public void checkingTwoVariablesAreNotEquals(String firstVariableName, String secondVariableName) {
-        String firstValueToCompare = akitaScenario.getVar(firstVariableName).toString();
-        String secondValueToCompare = akitaScenario.getVar(secondVariableName).toString();
-        assertThat(String.format("Значения в переменных [%s] и [%s] совпадают", firstVariableName, secondVariableName),
-            firstValueToCompare, Matchers.not(equalTo(secondValueToCompare)));
-    }
-
-    /**
-     * Проверка того, что значение из поля совпадает со значением заданной переменной из хранилища
-     */
-    @Тогда("^значение (?:поля|элемента) \"([^\"]*)\" совпадает со значением из переменной \"([^\"]*)\"$")
-    public void compareFieldAndVariable(String elementName, String variableName) {
-        String actualValue = akitaScenario.getCurrentPage().getAnyElementText(elementName);
-        String expectedValue = akitaScenario.getVar(variableName).toString();
-        assertThat(String.format("Значение поля [%s] не совпадает со значением из переменной [%s]", elementName, variableName),
-            actualValue, equalTo(expectedValue));
-    }
-
-    /**
-     * Проверка того, что значение из поля содержится в списке,
-     * полученном из хранилища переменных по заданному ключу
-     */
-    @SuppressWarnings("unchecked")
-    @Тогда("^список из переменной \"([^\"]*)\" содержит значение (?:поля|элемента) \"([^\"]*)\"$")
-    public void checkIfListContainsValueFromField(String variableListName, String elementName) {
-        String actualValue = akitaScenario.getCurrentPage().getAnyElementText(elementName);
-        List<String> listFromVariable = ((List<String>) akitaScenario.getVar(variableListName));
-        assertTrue(String.format("Список из переменной [%s] не содержит значение поля [%s]", variableListName, elementName),
-            listFromVariable.contains(actualValue));
-    }
-
-    /**
-     * Выполняется переход по заданной ссылке.
-     * Шаг содержит проверку, что после перехода загружена заданная страница.
-     * Ссылка может передаваться как строка, так и как ключ из application.properties
-     */
-    @И("^совершен переход на страницу \"([^\"]*)\" по ссылке \"([^\"]*)\"$")
-    public void goToSelectedPageByLink(String pageName, String urlOrName) {
-        String address = loadProperty(urlOrName, resolveVars(urlOrName));
-        akitaScenario.write(" url = " + address);
-        open(address);
-        loadPage(pageName);
-    }
-
-    /**
-     * Ожидание в течение заданного количества секунд
-     */
-    @Когда("^выполнено ожидание в течение (\\d+) (?:секунд|секунды)")
-    public void waitForSeconds(long seconds) {
-        sleep(1000 * seconds);
-    }
-
-    /**
-     * Проверка того, что блок исчез/стал невидимым
-     */
-    @Тогда("^(?:страница|блок|форма) \"([^\"]*)\" (?:скрыт|скрыта)")
-    public void blockDisappeared(String nameOfPage) {
-        if (isIE()) {
-            akitaScenario.getPage(nameOfPage).ieDisappeared();
-        } else akitaScenario.getPage(nameOfPage).disappeared();
-    }
-
-    /**
-     * Эмулирует нажатие клавиш на клавиатуре
-     */
-    @И("^выполнено нажатие на клавиатуре \"([^\"]*)\"$")
-    public void pushButtonOnKeyboard(String buttonName) {
-        Keys key = Keys.valueOf(buttonName.toUpperCase());
-        switchTo().activeElement().sendKeys(key);
-    }
-
-    /**
-     * Эмулирует нажатие сочетания клавиш на клавиатуре.
-     * Допустим, чтобы эмулировать нажатие на Ctrl+A, в таблице должны быть следующие значения
-     * | CONTROL |
-     * | a       |
-     *
-     * @param keyNames название клавиши
-     */
-    @И("^выполнено нажатие на сочетание клавиш из таблицы$")
-    public void pressKeyCombination(List<String> keyNames) {
-        Iterable<CharSequence> listKeys = keyNames.stream()
-            .map(this::getKeyOrCharacter)
-            .collect(Collectors.toList());
-        String combination = Keys.chord(listKeys);
-        switchTo().activeElement().sendKeys(combination);
-    }
-
-    private CharSequence getKeyOrCharacter(String key) {
-        try {
-            return Keys.valueOf(key.toUpperCase());
-        } catch (IllegalArgumentException ex) {
-            return key;
-        }
-    }
-
-    /**
-     * Устанавливается значение (в приоритете: из property, из переменной сценария, значение аргумента) в заданное поле.
-     * Перед использованием поле нужно очистить
-     */
-    @Когда("^в поле \"([^\"]*)\" введено значение \"(.*)\"$")
-    public void setFieldValue(String elementName, String value) {
-        value = getPropertyOrStringVariableOrValue(value);
-        SelenideElement valueInput = akitaScenario.getCurrentPage().getElement(elementName);
-        cleanField(elementName);
-        valueInput.setValue(value);
-    }
-
-    /**
-     * Очищается заданное поле
-     */
-    @Когда("^очищено поле \"([^\"]*)\"$")
-    public void cleanField(String nameOfField) {
-        SelenideElement valueInput = akitaScenario.getCurrentPage().getElement(nameOfField);
-        Keys removeKey = isIE() ? Keys.BACK_SPACE : Keys.DELETE;
-        do {
-            valueInput.doubleClick().sendKeys(removeKey);
-        } while (valueInput.getValue().length() != 0);
-    }
-
-    /**
-     * Проверка, что поле для ввода пусто
-     */
-    @Тогда("^поле \"([^\"]*)\" пусто$")
-    public void fieldInputIsEmpty(String fieldName) {
-        assertThat(String.format("Поле [%s] не пусто", fieldName),
-            akitaScenario.getCurrentPage().getAnyElementText(fieldName),
-            isEmptyOrNullString());
-    }
-
-    /**
-     * Устанавливает размеры окна браузера
-     */
-    @И("^установлено разрешение экрана (\\d+) х (\\d+)$")
-    public void setBrowserWindowSize(int width, int height) {
-        getWebDriver().manage().window().setSize(new Dimension(width, height));
-        akitaScenario.write("Установлены размеры окна браузера: ширина " + width + " высота" + height);
-    }
-
-    /**
-     * Разворачивает окно с браузером на весь экран
-     */
-    @Если("^окно развернуто на весь экран$")
-    public void expandWindowToFullScreen() {
-        getWebDriver().manage().window().maximize();
-    }
-
-    /**
-     * Проверка, что список со страницы состоит только из элементов,
-     * перечисленных в таблице
-     * Для получения текста из элементов списка используется метод getText()
-     */
-    @Тогда("^список \"([^\"]*)\" состоит из элементов из таблицы$")
-    public void checkIfListConsistsOfTableElements(String listName, List<String> textTable) {
-        List<String> actualValues = akitaScenario.getCurrentPage().getAnyElementsListTexts(listName);
-        int numberOfTypes = actualValues.size();
-        assertThat(String.format("Количество элементов в списке [%s] не соответсвует ожиданию", listName), textTable, hasSize(numberOfTypes));
-        assertTrue(String.format("Значения элементов в списке [%s] не совпадают с ожидаемыми значениями из таблицы", listName), actualValues.containsAll(textTable));
-    }
-
-    /**
-     * Проверка, что список со страницы состоит только из элементов,
-     * перечисленных в таблице
-     * Для получения текста из элементов списка используется метод innerText()
-     */
-    @Тогда("^список \"([^\"]*)\" состоит из элементов таблицы$")
-    public void checkIfListInnerTextConsistsOfTableElements(String listName, List<String> textTable) {
-        List<String> actualValues = akitaScenario.getCurrentPage().getAnyElementsListInnerTexts(listName);
-        int numberOfTypes = actualValues.size();
-        assertThat(String.format("Количество элементов в списке [%s] не соответсвует ожиданию", listName), textTable, hasSize(numberOfTypes));
-        assertTrue(String.format("Значения элементов в списке %s: %s не совпадают с ожидаемыми значениями из таблицы %s", listName, actualValues, textTable),
-            actualValues.containsAll(textTable));
-    }
-
-    /**
-     * Выбор из списка со страницы элемента с заданным значением
-     * (в приоритете: из property, из переменной сценария, значение аргумента)
-     */
-    @Тогда("^в списке \"([^\"]*)\" выбран элемент с (?:текстом|значением) \"(.*)\"$")
-    public void checkIfSelectedListElementMatchesValue(String listName, String expectedValue) {
-        final String value = getPropertyOrStringVariableOrValue(expectedValue);
-        List<SelenideElement> listOfElementsFromPage = akitaScenario.getCurrentPage().getElementsList(listName);
-        List<String> elementsText = listOfElementsFromPage.stream()
-            .map(element -> element.getText().trim())
-            .collect(toList());
-        listOfElementsFromPage.stream()
-            .filter(element -> element.getText().trim().equalsIgnoreCase(value))
-            .findFirst()
-            .orElseThrow(() -> new IllegalArgumentException(String.format("Элемент [%s] не найден в списке %s: [%s] ", value, listName, elementsText)))
-            .click();
-    }
-
-    /**
-     * Выбор из списка со страницы элемента, который содержит заданный текст
-     * (в приоритете: из property, из переменной сценария, значение аргумента)
-     * Не чувствителен к регистру
-     */
-    @Тогда("^в списке \"([^\"]*)\" выбран элемент содержащий текст \"([^\"]*)\"$")
-    public void selectElementInListIfFoundByText(String listName, String expectedValue) {
-        final String value = getPropertyOrStringVariableOrValue(expectedValue);
-        List<SelenideElement> listOfElementsFromPage = akitaScenario.getCurrentPage().getElementsList(listName);
-        List<String> elementsListText = listOfElementsFromPage.stream()
-            .map(element -> element.getText().trim().toLowerCase())
-            .collect(toList());
-        listOfElementsFromPage.stream()
-            .filter(element -> element.getText().trim().toLowerCase().contains(value.toLowerCase()))
-            .findFirst()
-            .orElseThrow(() -> new IllegalArgumentException(String.format("Элемент [%s] не найден в списке %s: [%s] ", value, listName, elementsListText)))
-            .click();
-    }
-
-    /**
-     * Проверка, что список со страницы совпадает со списком из переменной
-     * без учёта порядка элементов
-     * Для получения текста из элементов списка используется метод innerText()
-     */
-    @SuppressWarnings("unchecked")
-    @Тогда("^список \"([^\"]*)\" на странице совпадает со списком \"([^\"]*)\"$")
-    public void checkListInnerTextCorrespondsToListFromVariable(String listName, String listVariable) {
-        List<String> expectedList = new ArrayList<>((List<String>) akitaScenario.getVar(listVariable));
-        List<String> actualList = new ArrayList<>(akitaScenario.getCurrentPage().getAnyElementsListInnerTexts(listName));
-        assertThat(String.format("Количество элементов списка %s = %s, ожидаемое значение = %s", listName, actualList.size(), expectedList.size()), actualList,
-            hasSize(expectedList.size()));
-        assertThat(String.format("Список со страницы %s: %s не совпадает с ожидаемым списком из переменной %s:%s", listName, actualList, listVariable, expectedList)
-            , actualList, containsInAnyOrder(expectedList.toArray()));
-    }
-
-    /**
-     * Сохранение значения элемента в переменную
-     */
-    @Когда("^значение (?:элемента|поля) \"([^\"]*)\" сохранено в переменную \"([^\"]*)\"$")
-    public void storeElementValueInVariable(String elementName, String variableName) {
-        akitaScenario.setVar(variableName, akitaScenario.getCurrentPage().getAnyElementText(elementName));
-        akitaScenario.write("Значение [" + akitaScenario.getCurrentPage().getAnyElementText(elementName) + "] сохранено в переменную [" + variableName + "]");
-    }
-
-    /**
-     * Проверка выражения на истинность
-     * выражение из property, из переменной сценария или значение аргумента
-     * Например, string1.equals(string2)
-     * OR string.equals("string")
-     * Любое Java-выражение, возвращающие boolean
-     */
-    @Тогда("^верно, что \"([^\"]*)\"$")
-    public void expressionExpression(String expression) {
-        akitaScenario.getVars().evaluate("assert(" + expression + ")");
-    }
-
-    /**
-     * Переход на страницу по клику и проверка, что страница загружена
-     */
-    @И("^выполнен переход на страницу \"([^\"]*)\" после нажатия на (?:ссылку|кнопку) \"([^\"]*)\"$")
-    public void urlClickAndCheckRedirection(String pageName, String elementName) {
-        akitaScenario.getCurrentPage().getElement(elementName).click();
-        loadPage(pageName);
-        akitaScenario.write(" url = " + url());
-    }
-
-    /**
-     * Шаг авторизации.
-     * Для того, чтобы шаг работал, на текущей странице должны быть указаны элементы
-     * со значениями аннотации @Name:
-     * "Логин" - для поля ввода логина,
-     * "Пароль" - для поля ввода пароля и
-     * "Войти" - для кнопки входа.
-     * Также должны быть указаны логин и пароль в файле application.properties.
-     * Например для шага: "Пусть пользователь user ввел логин и пароль"
-     * логин и пароль должны быть указаны со следующими ключами:
-     * user.login - для логина и
-     * user.password - для пароля
-     */
-    @Пусть("^пользователь \"([^\"]*)\" ввел логин и пароль$")
-    public void loginByUserData(String userCode) {
-        String login = loadProperty(userCode + ".login");
-        String password = loadProperty(userCode + ".password");
-        cleanField("Логин");
-        akitaScenario.getCurrentPage().getElement("Логин").sendKeys(login);
-        cleanField("Пароль");
-        akitaScenario.getCurrentPage().getElement("Пароль").sendKeys(password);
-        akitaScenario.getCurrentPage().getElement("Войти").click();
-    }
-
-    /**
-     * Выполняется наведение курсора на элемент
-     */
-    @Когда("^выполнен ховер на (?:поле|элемент) \"([^\"]*)\"$")
-    public void elementHover(String elementName) {
-        SelenideElement field = akitaScenario.getCurrentPage().getElement(elementName);
-        field.hover();
-    }
-
-    /**
-     * Проверка того, что элемент не отображается на странице
-     */
-    @Тогда("^(?:поле|выпадающий список|элемент) \"([^\"]*)\" не отображается на странице$")
-    public void elementIsNotVisible(String elementName) {
-        akitaScenario.getCurrentPage().waitElementsUntil(
-            not(Condition.appear), DEFAULT_TIMEOUT, akitaScenario.getCurrentPage().getElement(elementName)
-        );
-    }
-
-    /**
-     * Проверка, что элемент на странице кликабелен
-     */
-    @Тогда("^(?:поле|элемент) \"([^\"]*)\" кликабельно$")
-    public void clickableField(String elementName) {
-        SelenideElement element = akitaScenario.getCurrentPage().getElement(elementName);
-        assertTrue(String.format("Элемент [%s] не кликабелен", elementName), element.isEnabled());
-    }
-
-    /**
-     * Проверка, что у элемента есть атрибут с ожидаемым значением (в приоритете: из property, из переменной сценария, значение аргумента)
-     */
-    @Тогда("^элемент \"([^\"]*)\" содержит атрибут \"([^\"]*)\" со значением \"(.*)\"$")
-    public void checkElemContainsAtrWithValue(String elementName, String attribute, String expectedAttributeValue) {
-        expectedAttributeValue = getPropertyOrStringVariableOrValue(expectedAttributeValue);
-        SelenideElement currentElement = akitaScenario.getCurrentPage().getElement(elementName);
-        String currentAtrValue = currentElement.attr(attribute);
-        assertThat(String.format("Элемент [%s] не содержит атрибут [%s] со значением [%s]", elementName, attribute, expectedAttributeValue)
-            , currentAtrValue, equalToIgnoringCase(expectedAttributeValue));
-    }
-
-    /**
-     * Проверка, что элемент содержит указанный класс (в приоритете: из property, из переменной сценария, значение аргумента)
-     * Например:
-     * если нужно проверить что элемент не отображается на странице, но проверки Selenium отрабатывают неверно,
-     * можно использовать данный метод и проверить, что среди его классов есть disabled
-     */
-    @Тогда("^элемент \"([^\"]*)\" содержит класс со значением \"(.*)\"$")
-    public void checkElemClassContainsExpectedValue(String elementName, String expectedClassValue) {
-        SelenideElement currentElement = akitaScenario.getCurrentPage().getElement(elementName);
-        expectedClassValue = getPropertyOrStringVariableOrValue(expectedClassValue);
-        String currentClassValue = currentElement.getAttribute("class");
-        assertThat(String.format("Элемент [%s] не содержит класс со значением [%s]", elementName, expectedClassValue)
-            , currentClassValue.toLowerCase(), containsString(expectedClassValue.toLowerCase()));
-    }
-
-    /**
-     * Проверка, что элемент не содержит указанный класс
-     */
-    @Тогда("^элемент \"([^\"]*)\" не содержит класс со значением \"(.*)\"$")
-    public void checkElemClassNotContainsExpectedValue(String elementName, String expectedClassValue) {
-        SelenideElement currentElement = akitaScenario.getCurrentPage().getElement(elementName);
-        assertThat(String.format("Элемент [%s] содержит класс со значением [%s]", elementName, expectedClassValue),
-            currentElement.getAttribute("class").toLowerCase(),
-            Matchers.not(containsString(getPropertyOrStringVariableOrValue(expectedClassValue).toLowerCase())));
-    }
-
-    /**
-     * Выполняется переход в конец страницы
-     */
-    @И("^совершен переход в конец страницы$")
-    public void scrollDown() {
-        Actions actions = new Actions(getWebDriver());
-        actions.keyDown(Keys.CONTROL).sendKeys(Keys.END).build().perform();
-        actions.keyUp(Keys.CONTROL).perform();
-    }
-
-    /**
-     * Проверка, что значение в поле содержит значение (в приоритете: из property, из переменной сценария, значение аргумента),
-     * указанное в шаге
-     */
-    @Тогда("^(?:поле|элемент) \"([^\"]*)\" содержит значение \"(.*)\"$")
-    public void testActualValueContainsSubstring(String elementName, String expectedValue) {
-        expectedValue = getPropertyOrStringVariableOrValue(expectedValue);
-        String actualValue = akitaScenario.getCurrentPage().getAnyElementText(elementName);
-        assertThat(String.format("Поле [%s] не содержит значение [%s]", elementName, expectedValue), actualValue, containsString(expectedValue));
-    }
-
-    /**
-     * Проверка, что значение в поле содержит текст, указанный в шаге
-     * (в приоритете: из property, из переменной сценария, значение аргумента).
-     * Используется метод innerText(), который получает как видимый, так и скрытый текст из элемента,
-     * обрезая перенос строк и пробелы в конце и начале строчки.
-     * Не чувствителен к регистру
-     */
-    @Тогда("^(?:поле|элемент) \"([^\"]*)\" содержит внутренний текст \"(.*)\"$")
-    public void testFieldContainsInnerText(String fieldName, String expectedText) {
-        expectedText = getPropertyOrStringVariableOrValue(expectedText);
-        String field = akitaScenario.getCurrentPage().getElement(fieldName).innerText().trim().toLowerCase();
-        assertThat(String.format("Поле [%s] не содержит текст [%s]", fieldName, expectedText), field, containsString(expectedText.toLowerCase()));
-    }
-
-    /**
-     * Проверка, что значение в поле равно значению, указанному в шаге (в приоритете: из property, из переменной сценария, значение аргумента)
-     */
-    @Тогда("^значение (?:поля|элемента) \"([^\"]*)\" равно \"(.*)\"$")
-    public void compareValInFieldAndFromStep(String elementName, String expectedValue) {
-        expectedValue = getPropertyOrStringVariableOrValue(expectedValue);
-        String actualValue = akitaScenario.getCurrentPage().getAnyElementText(elementName);
-        assertThat(String.format("Значение поля [%s] не равно ожидаемому [%s]", elementName, expectedValue), actualValue, equalTo(expectedValue));
-    }
-
-    /**
-     * Проверка, что кнопка/ссылка недоступна для нажатия
-     */
-    @Тогда("^(?:ссылка|кнопка) \"([^\"]*)\" недоступна для нажатия$")
-    public void buttonIsNotActive(String elementName) {
-        SelenideElement element = akitaScenario.getCurrentPage().getElement(elementName);
-        assertTrue(String.format("Элемент [%s] кликабелен", elementName), element.is(Condition.disabled));
-    }
-
-    /**
-     * Проверка, что поле нередактируемо
-     */
-    @Тогда("^(?:поле|элемент) \"([^\"]*)\" (?:недоступно|недоступен) для редактирования$")
-    public void fieldIsDisable(String elementName) {
-        SelenideElement element = akitaScenario.getCurrentPage().getElement(elementName);
-        assertTrue(String.format("Элемент [%s] доступен для редактирования", elementName), element.is(Condition.disabled));
-    }
-
-    /**
-     * Проверка, что список со страницы совпадает со списком из переменной
-     * без учёта порядка элементов
-     */
-    @SuppressWarnings("unchecked")
-    @Тогда("^список \"([^\"]*)\" со страницы совпадает со списком \"([^\"]*)\"$")
-    public void compareListFromUIAndFromVariable(String listName, String listVariable) {
-        HashSet<String> expectedList = new HashSet<>((List<String>) akitaScenario.getVar(listVariable));
-        HashSet<String> actualList = new HashSet<>(akitaScenario.getCurrentPage().getAnyElementsListTexts(listName));
-        assertThat(String.format("Список со страницы [%s] не совпадает с ожидаемым списком из переменной [%s]", listName, listVariable), actualList, equalTo(expectedList));
-    }
-
-    /**
-     * Проверка, что на странице не отображаются редактируемые элементы, такие как:
-     * -input
-     * -textarea
-     */
-    @Тогда("^открыта read-only форма$")
-    public void openReadOnlyForm() {
-        int inputsCount = getDisplayedElementsByCss("input").size();
-        assertTrue("Форма не read-only. Количество input-полей: " + inputsCount, inputsCount == 0);
-        int textareasCount = getDisplayedElementsByCss("textarea").size();
-        assertTrue("Форма не read-only. Количество элементов textarea: " + textareasCount, textareasCount == 0);
-    }
-
-    private List<SelenideElement> getDisplayedElementsByCss(String cssSelector) {
-        return $$(cssSelector).stream()
-            .filter(SelenideElement::isDisplayed)
-            .collect(Collectors.toList());
-    }
-
-    /**
-     * Добавление строки (в приоритете: из property, из переменной сценария, значение аргумента) в поле к уже заполненой строке
-     */
-    @Когда("^в элемент \"([^\"]*)\" дописывается значение \"(.*)\"$")
-    public void addValue(String elementName, String value) {
-        value = getPropertyOrStringVariableOrValue(value);
-        SelenideElement field = akitaScenario.getCurrentPage().getElement(elementName);
-        String oldValue = field.getValue();
-        if (oldValue.isEmpty()) {
-            oldValue = field.getText();
-        }
-        field.setValue("");
-        field.setValue(oldValue + value);
-    }
-
-    /**
-     * Нажатие на элемент по его тексту (в приоритете: из property, из переменной сценария, значение аргумента)
-     */
-    @И("^выполнено нажатие на элемент с текстом \"(.*)\"$")
-    public void findElement(String text) {
-        $(By.xpath(getTranslateNormalizeSpaceText(getPropertyOrStringVariableOrValue(text)))).click();
-    }
-
-    /**
-     * Ввод в поле текущей даты в заданном формате
-     * При неверном формате, используется dd.MM.yyyy
-     */
-    @Когда("^элемент \"([^\"]*)\" заполняется текущей датой в формате \"([^\"]*)\"$")
-    public void currentDate(String fieldName, String dateFormat) {
-        long date = System.currentTimeMillis();
-        String currentStringDate;
-        try {
-            currentStringDate = new SimpleDateFormat(dateFormat).format(date);
-        } catch (IllegalArgumentException ex) {
-            currentStringDate = new SimpleDateFormat("dd.MM.yyyy").format(date);
-            log.error("Неверный формат даты. Будет использоваться значание по умолчанию в формате dd.MM.yyyy");
-        }
-        SelenideElement valueInput = akitaScenario.getCurrentPage().getElement(fieldName);
-        valueInput.setValue("");
-        valueInput.setValue(currentStringDate);
-        akitaScenario.write("Текущая дата " + currentStringDate);
-    }
-
-    /**
-     * Ввод в поле указанного текста (в приоритете: из property, из переменной сценария, значение аргумента),
-     * используя буфер обмена и клавиши SHIFT + INSERT
-     */
-    @Когда("^вставлено значение \"([^\"]*)\" в элемент \"([^\"]*)\" с помощью горячих клавиш$")
-    public void pasteValueToTextField(String value, String fieldName) {
-        value = getPropertyOrStringVariableOrValue(value);
-        ClipboardOwner clipboardOwner = (clipboard, contents) -> {
-        };
-        Clipboard clipboard = Toolkit.getDefaultToolkit().getSystemClipboard();
-        StringSelection stringSelection = new StringSelection(value);
-        clipboard.setContents(stringSelection, clipboardOwner);
-        akitaScenario.getCurrentPage().getElement(fieldName).sendKeys(Keys.chord(Keys.SHIFT, Keys.INSERT));
-    }
-
-    /**
-     * Выполняется поиск нужного файла в папке /Downloads
-     * Поиск осуществляется по содержанию ожидаемого текста в названии файла. Можно передавать регулярное выражение.
-     * После выполнения проверки файл удаляется
-     */
-    @Тогда("^файл \"(.*)\" загрузился в папку /Downloads$")
-    public void testFileDownloaded(String fileName) {
-        File downloads = getDownloadsDir();
-        File[] expectedFiles = downloads.listFiles((files, file) -> file.contains(fileName));
-        assertNotNull("Ошибка поиска файла", expectedFiles);
-        assertFalse("Файл не загрузился", expectedFiles.length == 0);
-        assertTrue(String.format("В папке присутствуют более одного файла с одинаковым названием, содержащим текст [%s]", fileName),
-            expectedFiles.length == 1);
-        deleteFiles(expectedFiles);
-    }
-
-    /**
-     * Скроллит экран до нужного элемента, имеющегося на странице, но видимого только в нижней/верхней части страницы.
-     */
-    @Тогда("^страница прокручена до элемента \"([^\"]*)\"")
-    public void scrollPageToElement(String elementName) {
-        akitaScenario.getCurrentPage().getElement(elementName).scrollTo();
-    }
-
-    /**
-     * Выбор из списка со страницы любого случайного элемента
-     */
-    @Тогда("^выбран любой элемент в списке \"([^\"]*)\"$")
-    public void selectRandomElementFromList(String listName) {
-        List<SelenideElement> listOfElementsFromPage = akitaScenario.getCurrentPage().getElementsList(listName);
-        listOfElementsFromPage.get(getRandom(listOfElementsFromPage.size()))
-            .shouldBe(Condition.visible).click();
-        akitaScenario.write("Выбран случайный элемент: " + listOfElementsFromPage);
-    }
-
-    /**
-     * Выбор из списка со страницы любого случайного элемента и сохранение его значения в переменную
-     */
-    @Когда("^выбран любой элемент из списка \"([^\"]*)\" и его значение сохранено в переменную \"([^\"]*)\"$")
-    public void selectRandomElementFromListAndSaveVar(String listName, String varName) {
-        List<SelenideElement> listOfElementsFromPage = akitaScenario.getCurrentPage().getElementsList(listName);
-        SelenideElement element = listOfElementsFromPage.get(getRandom(listOfElementsFromPage.size()));
-        element.shouldBe(Condition.visible).click();
-        akitaScenario.setVar(varName, akitaScenario.getCurrentPage().getAnyElementText(element).trim());
-        akitaScenario.write(String.format("Переменной [%s] присвоено значение [%s] из списка [%s]", varName,
-            akitaScenario.getVar(varName), listName));
-    }
-
-    /**
-     * Выбор n-го элемента из списка со страницы
-     * Нумерация элементов начинается с 1
-     */
-    @Тогда("^выбран (\\d+)-й элемент в списке \"([^\"]*)\"$")
-    public void selectElementNumberFromList(Integer elementNumber, String listName) {
-        List<SelenideElement> listOfElementsFromPage = akitaScenario.getCurrentPage().getElementsList(listName);
-        SelenideElement elementToSelect;
-        Integer selectedElementNumber = elementNumber - 1;
-        if (selectedElementNumber < 0 || selectedElementNumber >= listOfElementsFromPage.size()) {
-            throw new IndexOutOfBoundsException(
-                String.format("В списке %s нет элемента с номером %s. Количество элементов списка = %s",
-                    listName, elementNumber, listOfElementsFromPage.size()));
-        }
-        elementToSelect = listOfElementsFromPage.get(selectedElementNumber);
-        elementToSelect.shouldBe(Condition.visible).click();
-    }
-
-    /**
-     * Выбор n-го элемента из списка со страницы и сохранение его значения в переменную
-     * Нумерация элементов начинается с 1
-     */
-    @Тогда("^выбран (\\d+)-й элемент в списке \"([^\"]*)\" и его значение сохранено в переменную \"([^\"]*)\"$")
-    public void selectElementNumberFromListAndSaveToVar(Integer elementNumber, String listName, String varName) {
-        List<SelenideElement> listOfElementsFromPage = akitaScenario.getCurrentPage().getElementsList(listName);
-        SelenideElement elementToSelect;
-        Integer selectedElementNumber = elementNumber - 1;
-        if (selectedElementNumber < 0 || selectedElementNumber >= listOfElementsFromPage.size()) {
-            throw new IndexOutOfBoundsException(
-                    String.format("В списке %s нет элемента с номером %s. Количество элементов списка = %s",
-                            listName, elementNumber, listOfElementsFromPage.size()));
-        }
-        elementToSelect = listOfElementsFromPage.get(selectedElementNumber);
-        elementToSelect.shouldBe(Condition.visible).click();
-        akitaScenario.setVar(varName, akitaScenario.getCurrentPage().getAnyElementText(elementToSelect).trim());
-    }
-
-    /**
-     * Проверка, что каждый элемент списка содержит ожидаемый текст
-     * Не чувствителен к регистру
-     */
-    @Тогда("^элементы списка \"([^\"]*)\" содержат текст \"([^\"]*)\"$")
-    public void checkListElementsContainsText(String listName, String expectedValue) {
-        final String value = getPropertyOrValue(expectedValue);
-        List<SelenideElement> listOfElementsFromPage = akitaScenario.getCurrentPage().getElementsList(listName);
-        List<String> elementsListText = listOfElementsFromPage.stream()
-            .map(element -> element.getText().trim().toLowerCase())
-            .collect(toList());
-        assertTrue(String.format("Элемены списка %s: [%s] не содержат текст [%s] ", listName, elementsListText, value),
-            elementsListText.stream().allMatch(item -> item.contains(value.toLowerCase())));
-    }
-
-    /**
-     * Проверка, что каждый элемент списка не содержит ожидаемый текст
-     */
-    @Тогда("^элементы списка \"([^\"]*)\" не содержат текст \"([^\"]*)\"$")
-    public void checkListElementsNotContainsText(String listName, String expectedValue) {
-        final String value = getPropertyOrValue(expectedValue);
-        List<SelenideElement> listOfElementsFromPage = akitaScenario.getCurrentPage().getElementsList(listName);
-        List<String> elementsListText = listOfElementsFromPage.stream()
-            .map(element -> element.getText().trim().toLowerCase())
-            .collect(toList());
-        assertFalse(String.format("Элемены списка %s: [%s] содержат текст [%s] ", listName, elementsListText, value),
-            elementsListText.stream().allMatch(item -> item.contains(value.toLowerCase())));
-    }
-
-    /**
-     * Ввод в поле случайной последовательности латинских или кириллических букв задаваемой длины
-     */
-    @Когда("^в поле \"([^\"]*)\" введено (\\d+) случайных символов на (кириллице|латинице)$")
-    public void setRandomCharSequence(String elementName, int seqLength, String lang) {
-        SelenideElement valueInput = akitaScenario.getCurrentPage().getElement(elementName);
-        cleanField(elementName);
-
-        if (lang.equals("кириллице")) lang = "ru";
-        else lang = "en";
-        String charSeq = getRandCharSequence(seqLength, lang);
-        valueInput.setValue(charSeq);
-        akitaScenario.write("Строка случайных символов равна :" + charSeq);
-    }
-
-    /**
-     * Ввод в поле случайной последовательности латинских или кириллических букв задаваемой длины и сохранение этого значения в переменную
-     */
-    @Когда("^в поле \"([^\"]*)\" введено (\\d+) случайных символов на (кириллице|латинице) и сохранено в переменную \"([^\"]*)\"$")
-    public void setRandomCharSequenceAndSaveToVar(String elementName, int seqLength, String lang, String varName) {
-        SelenideElement valueInput = akitaScenario.getCurrentPage().getElement(elementName);
-        cleanField(elementName);
-
-        if (lang.equals("кириллице")) lang = "ru";
-        else lang = "en";
-        String charSeq = getRandCharSequence(seqLength, lang);
-        valueInput.setValue(charSeq);
-        akitaScenario.setVar(varName, charSeq);
-        akitaScenario.write("Строка случайных символов равна :" + charSeq);
-    }
-
-    /**
-     * Ввод в поле случайной последовательности цифр задаваемой длины
-     */
-    @Когда("^в поле \"([^\"]*)\" введено случайное число из (\\d+) (?:цифр|цифры)$")
-    public void inputRandomNumSequence(String elementName, int seqLength) {
-        SelenideElement valueInput = akitaScenario.getCurrentPage().getElement(elementName);
-        cleanField(elementName);
-        String numSeq = RandomStringUtils.randomNumeric(seqLength);
-        valueInput.setValue(numSeq);
-        akitaScenario.write(String.format("В поле [%s] введено значение [%s]", elementName, numSeq));
-    }
-
-    /**
-     * Ввод в поле случайной последовательности цифр задаваемой длины и сохранение этого значения в переменную
-     */
-    @Когда("^в поле \"([^\"]*)\" введено случайное число из (\\d+) (?:цифр|цифры) и сохранено в переменную \"([^\"]*)\"$")
-    public void inputAndSetRandomNumSequence(String elementName, int seqLength, String varName) {
-        SelenideElement valueInput = akitaScenario.getCurrentPage().getElement(elementName);
-        cleanField(elementName);
-        String numSeq = RandomStringUtils.randomNumeric(seqLength);
-        valueInput.setValue(numSeq);
-        akitaScenario.setVar(varName, numSeq);
-        akitaScenario.write(String.format("В поле [%s] введено значение [%s] и сохранено в переменную [%s]",
-            elementName, numSeq, varName));
-    }
-
-    /**
-     * Проход по списку и проверка текста у элемента на соответствие формату регулярного выражения
-     */
-    @И("элементы списка \"([^\"]*)\" соответствуют формату \"([^\"]*)\"$")
-    public void checkListTextsByRegExp(String listName, String pattern) {
-        akitaScenario.getCurrentPage().getElementsList(listName).forEach(element -> {
-            String str = akitaScenario.getCurrentPage().getAnyElementText(element);
-            assertTrue(format("Текст '%s' из списка '%s' не соответствует формату регулярного выражения", str, listName),
-                isTextMatches(str, pattern));
-        });
-    }
-
-    /**
-     * Выполняется запуск js-скрипта с указанием в js.executeScript его логики
-     * Скрипт можно передать как аргумент метода или значение из application.properties
-     */
-    @Когда("^выполнен js-скрипт \"([^\"]*)\"")
-    public void executeJsScript(String scriptName) {
-        String content = loadValueFromFileOrPropertyOrVariableOrDefault(scriptName);
-        Selenide.executeJavaScript(content);
-    }
-
-    /**
-     * Производится проверка количества символов в поле со значением, указанным в шаге
-     */
-    @Тогда("^в поле \"([^\"]*)\" содержится (\\d+) символов$")
-    public void checkFieldSymbolsCount(String element, int num) {
-        int length = akitaScenario.getCurrentPage().getAnyElementText(element).length();
-        assertEquals(String.format("Неверное количество символов. Ожидаемый результат: %s, текущий результат: %s", num, length), num, length);
-    }
-
-    /**
-     * Производится проверка соответствия числа элементов списка значению, указанному в шаге
-     */
-    @Тогда("^в списке \"([^\"]*)\" содержится (\\d+) (?:элемент|элементов|элемента)")
-    public void listContainsNumberOfElements(String listName, int quantity) {
-        List<SelenideElement> listOfElementsFromPage = akitaScenario.getCurrentPage().getElementsList(listName);
-        assertTrue(String.format("Число элементов в списке отличается от ожидаемого: %s", listOfElementsFromPage.size()), listOfElementsFromPage.size() == quantity);
-    }
-
-    /**
-     * Производится проверка соответствия числа элементов списка значению из property файла, из переменной сценария или указанному в шаге
-     */
-    @Тогда("^в списке \"([^\"]*)\" содержится количество элементов, равное значению из переменной \"([^\"]*)\"")
-    public void listContainsNumberFromVariable(String listName, String quantity) {
-        int numberOfElements = getCounterFromString(getPropertyOrStringVariableOrValue(quantity));
-        listContainsNumberOfElements(listName, numberOfElements);
-    }
-
-    /**
-     * Производится сопоставление числа элементов списка и значения, указанного в шаге
-     */
-    @Тогда("^в списке \"([^\"]*)\" содержится (более|менее) (\\d+) (?:элементов|элемента)")
-    public void listContainsMoreOrLessElements(String listName, String moreOrLess, int quantity) {
-        List<SelenideElement> listOfElementsFromPage = akitaScenario.getCurrentPage().getElementsList(listName);
-        if ("более".equals(moreOrLess)) {
-            assertTrue(String.format("Число элементов списка меньше ожидаемого: %s", listOfElementsFromPage.size()), listOfElementsFromPage.size() > quantity);
-        } else
-            assertTrue(String.format("Число элементов списка превышает ожидаемое: %s", listOfElementsFromPage.size()), listOfElementsFromPage.size() < quantity);
-
-    }
-
-    /**
-     * Ввод в поле случайного дробного числа в заданном диапазоне и формате с последующим сохранением этого значения в переменную
-     * Пример формата ввода: ###.##
-     */
-    @Когда("^в поле \"([^\"]*)\" введено случайное дробное число от (\\d+) до (\\d+) в формате \"([^\"])\" и сохранено в переменную \"([^\"]*)\"$")
-    public void setRandomNumSequenceWithIntAndFract(String fieldName, double valueFrom, double valueTo, String outputFormat, String saveToVariableName) {
-        double finalValue = ThreadLocalRandom.current().nextDouble(valueFrom, valueTo);
-        setFieldValue(fieldName, new DecimalFormat(outputFormat).format(finalValue));
-        akitaScenario.setVar(saveToVariableName, new DecimalFormat(outputFormat).format(finalValue));
-        akitaScenario.write(String.format("В поле [%s] введено значение [%s] и сохранено в переменную [%s]",
-                fieldName, new DecimalFormat(outputFormat).format(finalValue), saveToVariableName));
-    }
-
-    /**
-     * Ввод в поле случайного дробного числа в заданном диапазоне и формате
-     * Пример формата ввода: ###.##
-     */
-    @Когда("^в поле \"([^\"]*)\" введено случайное дробное число от (\\d+) до (\\d+) в формате \"([^\"])\"$")
-    public void inputRandomNumSequenceWithIntAndFract(String fieldName, double valueFrom, double valueTo, String outputFormat) {
-        double finalValue = ThreadLocalRandom.current().nextDouble(valueFrom, valueTo);
-        setFieldValue(fieldName, new DecimalFormat(outputFormat).format(finalValue));
-        akitaScenario.write(String.format("В поле [%s] введено значение [%s]", fieldName, new DecimalFormat(outputFormat).format(finalValue)));
-    }
-
-    /**
-     *  Скроллит страницу вниз до появления элемента каждую секунду.
-     *  Если достигнут футер страницы и элемент не найден - выбрасывается exception.
-     * Скроллит страницу вниз до появления элемента каждую секунду.
-     * Если достигнут футер страницы и элемент не найден - выбрасывается exception.
-     */
-    @И("^страница прокручена до появления элемента \"([^\"]*)\"$")
-    public void scrollWhileElemNotFoundOnPage(String elementName) {
-        SelenideElement el = null;
-        do {
-            el = akitaScenario.getCurrentPage().getElement(elementName);
-            if (el.exists()) {
-                break;
-            }
-            executeJavaScript("return window.scrollBy(0, 250);");
-            sleep(1000);
-        } while (!atBottom());
-        assertThat("Элемент " + elementName + " не найден", el.isDisplayed());
-    }
-
-    /**
-     * Скроллит страницу вниз до появления элемента с текстом из property файла, из переменной сценария или указанному в шаге каждую секунду.
-     * Если достигнут футер страницы и элемент не найден - выбрасывается exception.
-     */
-    @И("^страница прокручена до появления элемента с текстом \"([^\"]*)\"$")
-    public void scrollWhileElemWithTextNotFoundOnPage(String expectedValue) {
-        SelenideElement el = null;
-        do {
-            el = $(By.xpath(getTranslateNormalizeSpaceText(getPropertyOrStringVariableOrValue(expectedValue))));
-            if (el.exists()) {
-                break;
-            }
-            executeJavaScript("return window.scrollBy(0, 250);");
-            sleep(1000);
-        } while (!atBottom());
-        assertThat("Элемент с текстом " + expectedValue + " не найден", el.isDisplayed());
-    }
-
-    /**
-     * Метод осуществляет снятие скриншота и прикрепление его к cucumber отчету.
-     */
-    @И("^снят скриншот текущей страницы$")
-    public void takeScreenshot() {
-        final byte[] screenshot = ((TakesScreenshot) getWebDriver()).getScreenshotAs(OutputType.BYTES);
-        AkitaScenario.getInstance().getScenario().embed(screenshot, "image/png");
-    }
-
-    /*
-     * Проверка совпадения значения из переменной и значения из property
-     */
-    @Тогда("^значения из переменной \"([^\"]*)\" и из property файла \"([^\"]*)\" совпадают$")
-    public void checkIfValueFromVariableEqualPropertyVariable(String envVarible, String propertyVariable) {
-        assertThat("Переменные " + envVarible + " и " + propertyVariable + " не совпадают",
-            (String) akitaScenario.getVar(envVarible), equalToIgnoringCase(loadProperty(propertyVariable)));
-    }
-
-    /*
-     * Выполняется нажатие на кнопку и подгружается указанный файл
-     * Селектор кнопки должны быть строго на input элемента
-     * Можно указать путь до файла. Например, src/test/resources/example.pdf
-     */
-    @Когда("^выполнено нажатие на кнопку \"([^\"]*)\" и загружен файл \"([^\"]*)\"$")
-    public void clickOnButtonAndUploadFile(String buttonName, String fileName) {
-        String file = loadValueFromFileOrPropertyOrVariableOrDefault(fileName);
-        File attachmentFile = new File(file);
-        akitaScenario.getCurrentPage().getElement(buttonName).uploadFile(attachmentFile);
-    }
-
-    /*
-     * Выполняется чтение файла с шаблоном и заполнение его значениями из таблицы
-     */
-    @И("^шаблон \"([^\"]*)\" заполнен данными из таблицы и сохранён в переменную \"([^\"]*)\"$")
-    public void fillTemplate(String templateName, String varName, DataTable table) {
-        String template = loadValueFromFileOrPropertyOrVariableOrDefault(templateName);
-        boolean error = false;
-        for (List<String> list : table.raw()) {
-            String regexp = list.get(0);
-            String replacement = list.get(1);
-            if (template.contains(regexp)) {
-                template = template.replaceAll(regexp, replacement);
-            } else {
-                akitaScenario.write("В шаблоне не найден элемент " + regexp);
-                error = true;
-            }
-        }
-        if (error)
-            throw new RuntimeException("В шаблоне не найдены требуемые регулярные выражения");
-        akitaScenario.setVar(varName, template);
-    }
-
-    /**
-     * Возвращает значение из property файла, если отсутствует, то из пользовательских переменных,
-     * если и оно отсутствует, то возвращает значение переданной на вход переменной
-     *
-     * @return
-     */
-    public String getPropertyOrStringVariableOrValue(String propertyNameOrVariableNameOrValue) {
-        String propertyValue = tryLoadProperty(propertyNameOrVariableNameOrValue);
-        String variableValue = (String) akitaScenario.tryGetVar(propertyNameOrVariableNameOrValue);
-
-        boolean propertyCheck = checkResult(propertyValue, "Переменная " + propertyNameOrVariableNameOrValue + " из property файла");
-        boolean variableCheck = checkResult(variableValue, "Переменная сценария " + propertyNameOrVariableNameOrValue);
-
-        return propertyCheck ? propertyValue : (variableCheck ? variableValue : propertyNameOrVariableNameOrValue);
-    }
-
-    private boolean checkResult(String result, String message) {
-        if (isNull(result)) {
-            log.warn(message + " не найдена");
-            return false;
-        }
-        log.info(message + " = " + result);
-        akitaScenario.write(message + " = " + result);
-        return true;
-    }
-
-    /**
-     * Возвращает каталог "Downloads" в домашней директории
-     *
-     * @return
-     */
-    private File getDownloadsDir() {
-        String homeDir = System.getProperty("user.home");
-        return new File(homeDir + "/Downloads");
-    }
-
-    /**
-     * Удаляет файлы, переданные в метод
-     *
-     * @param filesToDelete массив файлов
-     */
-    private void deleteFiles(File[] filesToDelete) {
-        for (File file : filesToDelete) {
-            file.delete();
-        }
-    }
-
-    /**
-     * Возвращает случайное число от нуля до maxValueInRange
-     *
-     * @param maxValueInRange максимальная граница диапазона генерации случайных чисел
-     */
-    private int getRandom(int maxValueInRange) {
-        return (int) (Math.random() * maxValueInRange);
-    }
-
-    /**
-     * Возвращает последовательность случайных символов переданных алфавита и длины
-     * Принимает на вход варианты языков 'ru' и 'en'
-     * Для других входных параметров возвращает латинские символы (en)
-     */
-    public String getRandCharSequence(int length, String lang) {
-
-        StringBuilder builder = new StringBuilder();
-        for (int i = 0; i < length; i++) {
-            char symbol = charGenerator(lang);
-            builder.append(symbol);
-        }
-        return builder.toString();
-    }
-
-    /**
-     * Возвращает случайный символ переданного алфавита
-     */
-    private char charGenerator(String lang) {
-        Random random = new Random();
-        if (lang.equals("ru")) {
-            return (char) (1072 + random.nextInt(32));
-        } else {
-            return (char) (97 + random.nextInt(26));
-        }
-    }
-
-    /**
-     * Проверка на соответствие строки паттерну
-     */
-    public boolean isTextMatches(String str, String pattern) {
-        Pattern r = Pattern.compile(pattern);
-        Matcher m = r.matcher(str);
-        return m.matches();
-    }
-
-    /**
-     * Возвращает локатор для поиска по нормализованному(без учета регистра) тексту
-     */
-    public String getTranslateNormalizeSpaceText(String expectedText) {
-        StringBuilder text = new StringBuilder();
-        text.append("//*[contains(translate(normalize-space(text()), ");
-        text.append("'ABCDEFGHIJKLMNOPQRSTUVWXYZАБВГДЕЁЖЗИЙКЛМНОПРСТУФХЦЧШЩЪЫЬЭЮЯ', ");
-        text.append("'abcdefghijklmnopqrstuvwxyzабвгдеёжзийклмнопрстуфхчшщъыьэюя'), '");
-        text.append(expectedText.toLowerCase());
-        text.append("')]");
-        return text.toString();
-    }
-
-    /**
-     * Выдергиваем число из строки
-     */
-    private int getCounterFromString(String variableName) {
-        return Integer.parseInt(variableName.replaceAll("[^0-9]",""));
-    }
-
-    /**
-     * Удаляет все пробелы из текста
-     */
-    private String removeSpacesFromValue(String variableName) {
-        return variableName.replaceAll("\\s+", "");
-    }
-}
