+++ src/main/java/ru/alfabank/steps/InputInteractionSteps.java
@@ -1,0 +1,200 @@
+/**
+ * Copyright 2017 Alfa Laboratory
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package ru.alfabank.steps;
+
+import com.codeborne.selenide.SelenideElement;
+import cucumber.api.java.ru.Когда;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.RandomStringUtils;
+import org.openqa.selenium.Keys;
+import ru.alfabank.alfatest.cucumber.api.AkitaScenario;
+
+import java.awt.*;
+import java.awt.datatransfer.Clipboard;
+import java.awt.datatransfer.ClipboardOwner;
+import java.awt.datatransfer.StringSelection;
+import java.text.DecimalFormat;
+import java.text.SimpleDateFormat;
+import java.util.concurrent.ThreadLocalRandom;
+
+import static ru.alfabank.tests.core.helpers.PropertyLoader.loadPropertyInt;
+
+/**
+ * Шаги для тестирования полей ввода, доступные по умолчанию в каждом новом проекте
+ */
+
+
+@Slf4j
+public class InputInteractionSteps extends BaseMethods {
+
+    private AkitaScenario akitaScenario = AkitaScenario.getInstance();
+
+    private static final int DEFAULT_TIMEOUT = loadPropertyInt("waitingCustomElementsTimeout", 10000);
+
+    /**
+     * Устанавливается значение (в приоритете: из property, из переменной сценария, значение аргумента) в заданное поле.
+     * Перед использованием поле нужно очистить
+     */
+    @Когда("^в поле \"([^\"]*)\" введено значение \"(.*)\"$")
+    public void setFieldValue(String elementName, String value) {
+        value = getPropertyOrStringVariableOrValue(value);
+        SelenideElement valueInput = akitaScenario.getCurrentPage().getElement(elementName);
+        cleanField(elementName);
+        valueInput.setValue(value);
+    }
+
+    /**
+     * Очищается заданное поле
+     */
+    @Когда("^очищено поле \"([^\"]*)\"$")
+    public void cleaningField(String nameOfField) {
+        super.cleanField(nameOfField);
+    };
+
+    /**
+     * Добавление строки (в приоритете: из property, из переменной сценария, значение аргумента) в поле к уже заполненой строке
+     */
+    @Когда("^в элемент \"([^\"]*)\" дописывается значение \"(.*)\"$")
+    public void addValue(String elementName, String value) {
+        value = getPropertyOrStringVariableOrValue(value);
+        SelenideElement field = akitaScenario.getCurrentPage().getElement(elementName);
+        String oldValue = field.getValue();
+        if (oldValue.isEmpty()) {
+            oldValue = field.getText();
+        }
+        field.setValue("");
+        field.setValue(oldValue + value);
+    }
+
+    /**
+     * Ввод в поле текущей даты в заданном формате
+     * При неверном формате, используется dd.MM.yyyy
+     */
+    @Когда("^элемент \"([^\"]*)\" заполняется текущей датой в формате \"([^\"]*)\"$")
+    public void currentDate(String fieldName, String dateFormat) {
+        long date = System.currentTimeMillis();
+        String currentStringDate;
+        try {
+            currentStringDate = new SimpleDateFormat(dateFormat).format(date);
+        } catch (IllegalArgumentException ex) {
+            currentStringDate = new SimpleDateFormat("dd.MM.yyyy").format(date);
+            log.error("Неверный формат даты. Будет использоваться значание по умолчанию в формате dd.MM.yyyy");
+        }
+        SelenideElement valueInput = akitaScenario.getCurrentPage().getElement(fieldName);
+        valueInput.setValue("");
+        valueInput.setValue(currentStringDate);
+        akitaScenario.write("Текущая дата " + currentStringDate);
+    }
+
+    /**
+     * Ввод в поле указанного текста (в приоритете: из property, из переменной сценария, значение аргумента),
+     * используя буфер обмена и клавиши SHIFT + INSERT
+     */
+    @Когда("^вставлено значение \"([^\"]*)\" в элемент \"([^\"]*)\" с помощью горячих клавиш$")
+    public void pasteValueToTextField(String value, String fieldName) {
+        value = getPropertyOrStringVariableOrValue(value);
+        ClipboardOwner clipboardOwner = (clipboard, contents) -> {
+        };
+        Clipboard clipboard = Toolkit.getDefaultToolkit().getSystemClipboard();
+        StringSelection stringSelection = new StringSelection(value);
+        clipboard.setContents(stringSelection, clipboardOwner);
+        akitaScenario.getCurrentPage().getElement(fieldName).sendKeys(Keys.chord(Keys.SHIFT, Keys.INSERT));
+    }
+
+    /**
+     * Ввод в поле случайной последовательности латинских или кириллических букв задаваемой длины
+     */
+    @Когда("^в поле \"([^\"]*)\" введено (\\d+) случайных символов на (кириллице|латинице)$")
+    public void setRandomCharSequence(String elementName, int seqLength, String lang) {
+        SelenideElement valueInput = akitaScenario.getCurrentPage().getElement(elementName);
+        cleanField(elementName);
+
+        if (lang.equals("кириллице")) lang = "ru";
+        else lang = "en";
+        String charSeq = getRandCharSequence(seqLength, lang);
+        valueInput.setValue(charSeq);
+        akitaScenario.write("Строка случайных символов равна :" + charSeq);
+    }
+
+    /**
+     * Ввод в поле случайной последовательности латинских или кириллических букв задаваемой длины и сохранение этого значения в переменную
+     */
+    @Когда("^в поле \"([^\"]*)\" введено (\\d+) случайных символов на (кириллице|латинице) и сохранено в переменную \"([^\"]*)\"$")
+    public void setRandomCharSequenceAndSaveToVar(String elementName, int seqLength, String lang, String varName) {
+        SelenideElement valueInput = akitaScenario.getCurrentPage().getElement(elementName);
+        cleanField(elementName);
+
+        if (lang.equals("кириллице")) lang = "ru";
+        else lang = "en";
+        String charSeq = getRandCharSequence(seqLength, lang);
+        valueInput.setValue(charSeq);
+        akitaScenario.setVar(varName, charSeq);
+        akitaScenario.write("Строка случайных символов равна :" + charSeq);
+    }
+
+    /**
+     * Ввод в поле случайной последовательности цифр задаваемой длины
+     */
+    @Когда("^в поле \"([^\"]*)\" введено случайное число из (\\d+) (?:цифр|цифры)$")
+    public void inputRandomNumSequence(String elementName, int seqLength) {
+        SelenideElement valueInput = akitaScenario.getCurrentPage().getElement(elementName);
+        cleanField(elementName);
+        String numSeq = RandomStringUtils.randomNumeric(seqLength);
+        valueInput.setValue(numSeq);
+        akitaScenario.write(String.format("В поле [%s] введено значение [%s]", elementName, numSeq));
+    }
+
+    /**
+     * Ввод в поле случайной последовательности цифр задаваемой длины и сохранение этого значения в переменную
+     */
+    @Когда("^в поле \"([^\"]*)\" введено случайное число из (\\d+) (?:цифр|цифры) и сохранено в переменную \"([^\"]*)\"$")
+    public void inputAndSetRandomNumSequence(String elementName, int seqLength, String varName) {
+        SelenideElement valueInput = akitaScenario.getCurrentPage().getElement(elementName);
+        cleanField(elementName);
+        String numSeq = RandomStringUtils.randomNumeric(seqLength);
+        valueInput.setValue(numSeq);
+        akitaScenario.setVar(varName, numSeq);
+        akitaScenario.write(String.format("В поле [%s] введено значение [%s] и сохранено в переменную [%s]",
+                elementName, numSeq, varName));
+    }
+
+    /**
+     * Ввод в поле случайного дробного числа в заданном диапазоне и формате с последующим сохранением этого значения в переменную
+     * Пример формата ввода: ###.##
+     */
+    @Когда("^в поле \"([^\"]*)\" введено случайное дробное число от (\\d+) до (\\d+) в формате \"([^\"])\" и сохранено в переменную \"([^\"]*)\"$")
+    public void setRandomNumSequenceWithIntAndFract(String fieldName, double valueFrom, double valueTo, String outputFormat, String saveToVariableName) {
+        double finalValue = ThreadLocalRandom.current().nextDouble(valueFrom, valueTo);
+        setFieldValue(fieldName, new DecimalFormat(outputFormat).format(finalValue));
+        akitaScenario.setVar(saveToVariableName, new DecimalFormat(outputFormat).format(finalValue));
+        akitaScenario.write(String.format("В поле [%s] введено значение [%s] и сохранено в переменную [%s]",
+                fieldName, new DecimalFormat(outputFormat).format(finalValue), saveToVariableName));
+    }
+
+    /**
+     * Ввод в поле случайного дробного числа в заданном диапазоне и формате
+     * Пример формата ввода: ###.##
+     */
+    @Когда("^в поле \"([^\"]*)\" введено случайное дробное число от (\\d+) до (\\d+) в формате \"([^\"])\"$")
+    public void inputRandomNumSequenceWithIntAndFract(String fieldName, double valueFrom, double valueTo, String outputFormat) {
+        double finalValue = ThreadLocalRandom.current().nextDouble(valueFrom, valueTo);
+        setFieldValue(fieldName, new DecimalFormat(outputFormat).format(finalValue));
+        akitaScenario.write(String.format("В поле [%s] введено значение [%s]", fieldName, new DecimalFormat(outputFormat).format(finalValue)));
+    }
+
+
+}
