+++ src/main/java/ru/alfabank/steps/InputVerificationSteps.java
@@ -1,0 +1,48 @@
+/**
+ * Copyright 2017 Alfa Laboratory
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package ru.alfabank.steps;
+
+import cucumber.api.java.ru.Тогда;
+import lombok.extern.slf4j.Slf4j;
+import ru.alfabank.alfatest.cucumber.api.AkitaScenario;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.isEmptyOrNullString;
+import static ru.alfabank.tests.core.helpers.PropertyLoader.loadPropertyInt;
+
+/**
+ * Шаги для тестирования полей ввода, доступные по умолчанию в каждом новом проекте
+ */
+
+
+@Slf4j
+public class InputVerificationSteps extends BaseMethods {
+
+    private AkitaScenario akitaScenario = AkitaScenario.getInstance();
+
+    private static final int DEFAULT_TIMEOUT = loadPropertyInt("waitingCustomElementsTimeout", 10000);
+
+    /**
+     * Проверка, что поле для ввода пусто
+     */
+    @Тогда("^поле \"([^\"]*)\" пусто$")
+    public void fieldInputIsEmpty(String fieldName) {
+        assertThat(String.format("Поле [%s] не пусто", fieldName),
+                akitaScenario.getCurrentPage().getAnyElementText(fieldName),
+                isEmptyOrNullString());
+    }
+}
