--- src/test/java/ru/alfabank/steps/PageElementTest.java
@@ -1,78 +1,0 @@
-/**
- * Copyright 2017 Alfa Laboratory
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package ru.alfabank.steps;
-
-import com.codeborne.selenide.WebDriverRunner;
-import cucumber.api.Scenario;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import ru.alfabank.StubScenario;
-import ru.alfabank.alfatest.cucumber.api.AkitaEnvironment;
-import ru.alfabank.alfatest.cucumber.api.AkitaScenario;
-
-import java.io.File;
-
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
-
-public class PageElementTest {
-
-    private static DefaultSteps ds;
-    private static DefaultPageElementSteps dpes;
-    private static AkitaScenario akitaScenario;
-
-    @BeforeClass
-    public static void setup() {
-        akitaScenario = AkitaScenario.getInstance();
-        Scenario scenario = new StubScenario();
-        akitaScenario.setEnvironment(new AkitaEnvironment(scenario));
-        ds = new DefaultSteps();
-        dpes = new DefaultPageElementSteps();
-        String inputFilePath = "src/test/resources/AkitaPageMock.html";
-        String url = new File(inputFilePath).getAbsolutePath();
-        akitaScenario.setVar("Page", "file://" + url);
-    }
-
-    @Before
-    public void prepare() {
-        ds.goToSelectedPageByLink("AkitaPageMock", akitaScenario.getVar("Page").toString());
-    }
-
-    @AfterClass
-    public static void close() {
-        WebDriverRunner.closeWebDriver();
-    }
-
-    @Test
-    public void clickOnElementInBlockPositiveTest() {
-        dpes.clickOnElementInBlock("SearchButton", "SearchBlock");
-        assertFalse(akitaScenario.getPage("SearchBlock").getElement("SearchButton").isEnabled());
-    }
-
-    @Test
-    public void getElementsListInBlockPositiveTest() {
-        dpes.getElementsList("AkitaTable", "Rows", "ListTable");
-        assertNotNull(akitaScenario.getVar("ListTable"));
-    }
-
-    @Test
-    public void getListElementsTextInBlockPositiveTest() {
-        dpes.getListElementsText("AkitaTable", "Rows", "ListTable");
-        assertNotNull(akitaScenario.getVar("ListTable"));
-    }
-}
